# docs/swagger.yaml
basePath: /
definitions:
...
  models.PlayTypeDto:
    properties:
      is_active:
        type: boolean
      name:
        type: string
      promotion_time:
        type: integer
    type: object
...
  utils.StandardErrorResponse:
    properties:
      data:
        description: "null"
      message:
        example: Failed
        type: string
      success:
        example: false
        type: boolean
    type: object
  utils.StandardSuccessResponse:
    properties:
      data: {}
      message:
        example: Success
        type: string
      success:
        example: true
        type: boolean
    type: object
info:
  contact: {}
  description: This is a POS API with Gin and Swagger
  title: New POS APIls
  version: "1.0"
paths:
...
  /api/meter-record/all-spend/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Search member tel
        in: query
        name: memberTel
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get Meter Record all spend List
      tags:
      - Meter Record
  /api/meter-record/check-card/export:
    get:
      consumes:
      - application/json
      parameters:
      - description: Search card number
        in: query
        name: cardNo
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Export check card list
      tags:
      - Meter Record
  /api/meter-record/check-card/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Search card number
        in: query
        name: cardNo
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get Meter Record List with mobile number and start date
      tags:
      - Meter Record
  /api/meter-record/jubu-jibi:
    get:
      consumes:
      - application/json
      parameters:
      - description: Search member tel
        in: query
        name: memberTel
        type: string
      - description: Search location
        in: query
        name: location
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get Meter Record Jubu Jibi
      tags:
      - Meter Record
  /api/meter-record/jubu-jibi/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Search member tel
        in: query
        name: memberTel
        type: string
      - description: Search location
        in: query
        name: location
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get Meter Record Jubu Jibi List
      tags:
      - Meter Record
  /api/meter-record/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Search member tel
        in: query
        name: memberTel
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get Meter Record List with mobile number and start date
      tags:
      - Meter Record
  /api/meter-record/prize-status/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: Search member tel
        in: query
        name: memberTel
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Update Meter Record after refund request
      tags:
      - Meter Record
  /api/play-type:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get play type all active
      tags:
      - Play Type
    post:
      consumes:
      - application/json
      parameters:
      - description: Play Type Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.PlayTypeDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Create a new card type
      tags:
      - Play Type
  /api/play-type/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Delete play type by ID
      tags:
      - Play Type
    get:
      consumes:
      - application/json
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get play type by id
      tags:
      - Play Type
    put:
      consumes:
      - application/json
      parameters:
      - description: Play Type Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.PlayTypeDto'
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Update play type by ID
      tags:
      - Play Type
  /api/play-type/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Search text
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get play type list with paggination
      tags:
      - Play Type
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"